Move AWS code into brkt_cli.aws
Move 99% of the AWS-specific code into brkt_cli.aws.  Update
brkt_cli.main() to use the module interface to run registered commands.
Move the encrypt_ami() and update_ami() callsite to
AWSModuleInterface.

There's still some boto logging config code that's left in the brkt_cli
module.  This will get cleaned up in a subsequent commit.

Testing:
* python -m unittest discover
* run encrypt-ami manually
* run update-encrypted-ami manually
